requirements-dev.txt
black==19.3b0
factory-boy~=2.12.0
pytest~=5.0.1
pytest-django~=3.5.1
